fix: moved component where its used, removed unnecessary div
