+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/desktop/status-chip.tsx
@@ -2,8 +2,8 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import { colorPalette as lightModeColorPalette } from '@/shared/styles/color-palette';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { type MyJob } from '../../components';
-import { getChipStatusColor } from '../utils';
+import { type MyJob } from '../../../components';
+import { getChipStatusColor } from '../../utils';
 
 export function StatusChip({ status }: Readonly<{ status: MyJob['status'] }>) {
   const { colorPalette } = useColorMode();
